+++ monopyly/CHANGELOG.md
@@ -203,6 +203,7 @@
 
 - Style credit transaction submissions as receipts
 - Style flash messages according to content
+- Return to statement details page after deleting a transaction (rather than returning to the general transactions page)
 - Increase the flexibility of the credit activity parser